refactor(renderer): rename files
also, update comments
follow-up of #1064
