+++ pkg/renderer/sgml/html5/html5.go
@@ -8,8 +8,7 @@ import (
 	"github.com/bytesparadise/libasciidoc/pkg/types"
 )
 
-// Render renders the document to the output, using a default instance
-// of the renderer, with default templates.
+// Render renders the document to the output, using the SGML renderer configured with the HTML5 templates
 func Render(doc *types.Document, config *configuration.Configuration, output io.Writer) (types.Metadata, error) {
 	return sgml.Render(doc, config, output, templates)
 }
